revert template movement and const only for needed functions
