+++ platformio.ini
@@ -16,23 +16,24 @@ lib_deps =
 	; bodmer/TFT_eSPI@^2.5.43
 extra_scripts = 
 	pre:platformio_version_increment/version_increment_pre.py
+check_tool = cppcheck, clangtidy
 
 [env:esp32-c6-devkitc-1]
 platform = espressif32
 board = esp32-c6-devkitc-1
 
 [env:esp32-c6-devkitm-1]
-; platform = https://github.com/pioarduino/platform-espressif32/releases/download/51.03.05/platform-espressif32.zip
+platform = https://github.com/pioarduino/platform-espressif32/releases/download/51.03.05/platform-espressif32.zip
 ; platform = https://github.com/pioarduino/platform-espressif32/releases/download/51.03.07/platform-espressif32.zip
-platform = https://github.com/pioarduino/platform-espressif32/releases/download/stable/platform-espressif32.zip
+; platform = https://github.com/pioarduino/platform-espressif32/releases/download/stable/platform-espressif32.zip
 ; platform = espressif32
 board = esp32-c6-devkitm-1
-build_unflags = 
-;   -std=gnu++17
-	-std=gnu++2a
-	-std=gnu++2b
+; build_unflags = 
+; ;   -std=gnu++17
+; 	-std=gnu++2a
+; 	-std=gnu++2b
 build_flags =
-	-std=gnu++17
+	; -std=gnu++17
 ;   -std=gnu++2b
 	-D ARDUINO_USB_MODE=1
 	-D ARDUINO_USB_CDC_ON_BOOT=1
